Clean up middlewares
