+++ server/lib/middlewares/index.js
@@ -1,5 +1,4 @@
 export errorHandler from './errorHandler';
 export getUserAccessLevel from './getUserAccessLevel';
 export hasAccessLevel from './hasAccessLevel';
-export getStorage from './getStorage';
-export customMiddles from './customMiddles';
+export verifyUserAccess from './verifyUserAccess';
