+++ server/lib/middlewares/verifyUserAccess.js
@@ -1,0 +1,29 @@
+import { NotFoundError } from 'auth0-extension-tools';
+
+import * as constants from '../../constants';
+
+module.exports = (scriptManager) => (req, res, next) => {
+  if (req.user.role === constants.SUPER_ACCESS_LEVEL) {
+    return next();
+  }
+
+  return req.auth0.users.get({ id: req.params.id })
+    .then(user => {
+      if (!user) {
+        next(new NotFoundError(`User not found: ${req.params.id}`));
+      }
+
+      const accessContext = {
+        request: {
+          user: req.user
+        },
+        payload: {
+          user
+        }
+      };
+
+      return scriptManager.execute('access', accessContext);
+    })
+    .then(next)
+    .catch(next);
+};
